requirements.txt
h5py
matplotlib
numpy
numba
pillow